Fixed support for testnet3
Listed testnet3 magic bytes were wrong.
Bytes also need to be set before hand, was originally set to send
mainnet magicbytes in the first communication, testnet3 nodes wouldn't
reply to this.

Changed networks from list to magic bytes dict, previously the got_data
method would just check to see if the bytes were in the network list,
it wouldn't check to see if the message was from the network chosen.

Set network in settings, when sending or recieving messages the magic
bytes are read from the dict